deploy/AWS/Lambda/requirements.txt
onnxruntime
numpy
pillow
boto3